@@ -1,3 +1,3 @@
 module quesma/e2e-data-generator
 
-go 1.23.2
+go 1.24.0
