@@ -1,6 +1,6 @@
 module quesma.com/its
 
-go 1.23.2
+go 1.24.0
 
 require (
 	github.com/ClickHouse/clickhouse-go/v2 v2.20.0
