@@ -1,23 +1,24 @@
 module github.com/QuesmaOrg/quesma/quesma
 
-go 1.23.2
+go 1.24.0
 
 require (
-	github.com/ClickHouse/clickhouse-go/v2 v2.30.1
+	github.com/ClickHouse/clickhouse-go/v2 v2.32.2
 	github.com/DATA-DOG/go-sqlmock v1.5.2
-	github.com/DataDog/go-sqllexer v0.0.20
+	github.com/DataDog/go-sqllexer v0.0.21
 	github.com/H0llyW00dzZ/cidr v1.2.1
 	github.com/apparentlymart/go-cidr v1.1.0
 	github.com/barkimedes/go-deepcopy v0.0.0-20220514131651-17c30cfc62df
 	github.com/coreos/go-semver v0.3.1
-	github.com/go-sql-driver/mysql v1.8.1
-	github.com/goccy/go-json v0.10.4
-	github.com/google/go-cmp v0.6.0
+	github.com/go-sql-driver/mysql v1.9.0
+	github.com/goccy/go-json v0.10.5
+	github.com/google/go-cmp v0.7.0
 	github.com/google/uuid v1.6.0
 	github.com/gorilla/mux v1.8.1
 	github.com/gorilla/securecookie v1.1.2
 	github.com/gorilla/sessions v1.4.0
 	github.com/hashicorp/go-multierror v1.1.1
+	github.com/hashicorp/golang-lru/v2 v2.0.7
 	github.com/jackc/pgx/v5 v5.7.2
 	github.com/k0kubun/pp v3.0.1+incompatible
 	github.com/knadh/koanf/parsers/json v0.1.0
