@@ -1,5 +1,6 @@
 // Copyright Quesma, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
+
 // Experimental alpha processor for MySQL protocol
 
 package processors
@@ -13,19 +14,19 @@ import (
 	"net"
 )
 
-type TcpPassthroughMysqlProcessor struct {
+type TcpPassthroughMySqlProcessor struct {
 	BaseProcessor
 }
 
-func (t TcpPassthroughMysqlProcessor) InstanceName() string {
-	return "TcpPassthroughMysqlProcessor"
+func (t TcpPassthroughMySqlProcessor) InstanceName() string {
+	return "TcpPassthroughMySqlProcessor"
 }
 
-func (t TcpPassthroughMysqlProcessor) GetId() string {
-	return "TcpPassthroughMysqlProcessor"
+func (t TcpPassthroughMySqlProcessor) GetId() string {
+	return "TcpPassthroughMySqlProcessor"
 }
 
-func (t *TcpPassthroughMysqlProcessor) Handle(metadata map[string]interface{}, messages ...any) (map[string]interface{}, any, error) {
+func (t *TcpPassthroughMySqlProcessor) Handle(metadata map[string]interface{}, messages ...any) (map[string]interface{}, any, error) {
 	//
 	// MySQL client -> tcp_mysql_connection_handler -> tcp_passthrough_mysql_processor -> tcp_backend_connector -> real MySQL server
 	//
@@ -70,7 +71,7 @@ func (t *TcpPassthroughMysqlProcessor) Handle(metadata map[string]interface{}, m
 	//
 	// and forwards them back to the MySQL client.
 	for {
-		fullPacketBytes, err := frontend_connectors.ReadMysqlPacket(conn)
+		fullPacketBytes, err := frontend_connectors.ReadMySqlPacket(conn)
 		if err != nil {
 			break
 		}
@@ -109,8 +110,8 @@ func maybeProcessComQuery(msg []byte) []byte {
 	return msg
 }
 
-func NewTcpMysqlPassthroughProcessor() *TcpPassthroughMysqlProcessor {
-	return &TcpPassthroughMysqlProcessor{
+func NewTcpMySqlPassthroughProcessor() *TcpPassthroughMySqlProcessor {
+	return &TcpPassthroughMySqlProcessor{
 		BaseProcessor: NewBaseProcessor(),
 	}
 }
