@@ -6,11 +6,17 @@ package backend_connectors
 import (
 	"context"
 	"database/sql"
+	"github.com/QuesmaOrg/quesma/quesma/quesma/config"
 	quesma_api "github.com/QuesmaOrg/quesma/quesma/v2/core"
 )
 
+type SqlBackendConnector interface {
+	GetDB() *sql.DB
+}
+
 type BasicSqlBackendConnector struct {
 	connection *sql.DB
+	cfg        *config.RelationalDbConfiguration
 }
 
 type SqlRows struct {
@@ -34,12 +40,16 @@ func (p *SqlRows) Err() error {
 }
 
 func (p *BasicSqlBackendConnector) Open() error {
-	conn, err := initDBConnection()
+	conn, err := initDBConnection(p.cfg)
 	if err != nil {
 		return err
 	}
 	p.connection = conn
 	return nil
+}
+
+func (p *BasicSqlBackendConnector) GetDB() *sql.DB {
+	return p.connection
 }
 
 func (p *BasicSqlBackendConnector) Close() error {
