@@ -156,7 +156,13 @@ func BuildAttrsMap(m SchemaMap, config *ChTableConfig) (map[string][]interface{}
 			if a.Type.CanConvert(value) {
 				result[a.KeysArrayName] = append(result[a.KeysArrayName], name)
 				result[a.ValuesArrayName] = append(result[a.ValuesArrayName], fmt.Sprintf("%v", value))
-				result[a.TypesArrayName] = append(result[a.TypesArrayName], NewType(value).String())
+
+				valueType, err := NewType(value, name)
+				if err != nil {
+					result[a.TypesArrayName] = append(result[a.TypesArrayName], UndefinedType)
+				} else {
+					result[a.TypesArrayName] = append(result[a.TypesArrayName], valueType.String())
+				}
 
 				matched = true
 				break
