@@ -13,8 +13,8 @@ import (
 	"github.com/QuesmaOrg/quesma/quesma/elasticsearch/feature"
 	"github.com/QuesmaOrg/quesma/quesma/end_user_errors"
 	"github.com/QuesmaOrg/quesma/quesma/logger"
+	"github.com/QuesmaOrg/quesma/quesma/parsers/elastic_query_dsl"
 	"github.com/QuesmaOrg/quesma/quesma/processors/es_to_ch_common"
-	"github.com/QuesmaOrg/quesma/quesma/queryparser"
 	"github.com/QuesmaOrg/quesma/quesma/quesma/config"
 	"github.com/QuesmaOrg/quesma/quesma/quesma/recovery"
 	"github.com/QuesmaOrg/quesma/quesma/quesma/types"
@@ -34,7 +34,7 @@ import (
 
 func responseFromElastic(ctx context.Context, elkResponse *http.Response, w http.ResponseWriter) {
 	if id, ok := ctx.Value(tracing.RequestIdCtxKey).(string); ok {
-		logger.Debug().Str(logger.RID, id).Msg("responding from Elasticsearch")
+		logger.Debug().Str(logger.RID, id).Msgf("responding from Elasticsearch, status_code=%d", elkResponse.StatusCode)
 	}
 
 	copyHeaders(w, elkResponse)
@@ -149,8 +149,8 @@ func (*Dispatcher) closedIndexResponse(ctx context.Context, w http.ResponseWrite
 
 	response := make(types.JSON)
 
-	response["error"] = queryparser.Error{
-		RootCause: []queryparser.RootCause{
+	response["error"] = elastic_query_dsl.Error{
+		RootCause: []elastic_query_dsl.RootCause{
 			{
 				Type:   "index_closed_exception",
 				Reason: fmt.Sprintf("pattern %s is not routed to any connector", pattern),
@@ -196,7 +196,7 @@ func (r *Dispatcher) ElasticFallback(decision *quesma_api.Decision,
 			w.Header().Set(QuesmaSourceHeader, QuesmaSourceClickhouse)
 			AddProductAndContentHeaders(req.Header, w.Header())
 			w.WriteHeader(http.StatusNoContent)
-			w.Write(queryparser.EmptySearchResponse(ctx))
+			w.Write(elastic_query_dsl.EmptySearchResponse(ctx))
 			return
 		}
 
@@ -221,6 +221,7 @@ func (r *Dispatcher) ElasticFallback(decision *quesma_api.Decision,
 			}
 			feature.AnalyzeUnsupportedCalls(ctx, req.Method, req.URL.Path, req.Header.Get(OpaqueIdHeaderKey), resolveIndexPattern)
 		}
+		logger.DebugWithCtx(ctx).Msgf("request to path=%s routed to Elasticsearch", req.URL)
 		rawResponse := <-r.sendHttpRequestToElastic(ctx, req, reqBody, true)
 		response := rawResponse.response
 		if response != nil {
@@ -237,10 +238,10 @@ func (r *Dispatcher) ElasticFallback(decision *quesma_api.Decision,
 	}
 }
 
-func (r *Dispatcher) Reroute(ctx context.Context, w http.ResponseWriter, req *http.Request, reqBody []byte, router quesma_api.Router, logManager *clickhouse.LogManager, schemaRegistry schema.Registry) {
+func (r *Dispatcher) Reroute(ctx context.Context, w http.ResponseWriter, req *http.Request, reqBody []byte, router quesma_api.Router) {
 	defer recovery.LogAndHandlePanic(ctx, func(err error) {
 		w.WriteHeader(500)
-		w.Write(queryparser.InternalQuesmaError("Unknown Quesma error"))
+		w.Write(elastic_query_dsl.InternalQuesmaError("Unknown Quesma error"))
 	})
 
 	quesmaRequest, ctx, err := preprocessRequest(ctx, &quesma_api.Request{
@@ -273,7 +274,9 @@ func (r *Dispatcher) Reroute(ctx context.Context, w http.ResponseWriter, req *ht
 		quesmaResponse, err := recordRequestToClickhouse(req.URL.Path, r.debugInfoCollector, func() (*quesma_api.Result, error) {
 			var result *quesma_api.Result
 			result, err = handlersPipe.Handler(ctx, quesmaRequest, w)
-
+			if err != nil {
+				logger.ErrorWithCtx(ctx).Msgf("Error processing request: %v", err)
+			}
 			if result == nil {
 				return result, err
 			}
@@ -444,7 +447,9 @@ func PeekBodyV2(r *http.Request) ([]byte, error) {
 	reqBody, err := io.ReadAll(r.Body)
 	if err != nil {
 		logger.ErrorWithCtxAndReason(r.Context(), "incomplete request").
-			Msgf("Error reading request body: %v", err)
+			Msgf("Error reading request body: %v, url=%v", err, r.URL)
+		// To further examine erroneous body sent from the client
+		logger.DebugWithCtx(r.Context()).Msgf("Failed peek body | Bytes read=[%d] | Content length=[%d] | Headers=[%v]", len(reqBody), r.ContentLength, r.Header)
 		return nil, err
 	}
 
