@@ -77,7 +77,7 @@ func TestParseSignozSchema_2(t *testing.T) {
 		assert.Contains(t, table.Cols, fieldName)
 	}
 	assert.Equal(t, "db", table.DatabaseName)
-	assert.Equal(t, "cluster", table.Cluster)
+	assert.Equal(t, "cluster", table.ClusterName)
 }
 
 func TestParseQuotedTablename(t *testing.T) {
@@ -117,6 +117,79 @@ func TestParseNonLetterNames(t *testing.T) {
 	for _, fieldName := range fieldNames {
 		assert.Contains(t, table.Cols, fieldName)
 	}
+}
+
+func TestParseCreateSampleDataEcommerce(t *testing.T) {
+	q := `CREATE TABLE IF NOT EXISTS "kibana_sample_data_ecommerce" ON CLUSTER "quesma_cluster" 
+(
+	"@timestamp" DateTime64(3) DEFAULT now64(),
+	"attributes_values" Map(String,String),
+	"attributes_metadata" Map(String,String),
+
+
+	"taxful_total_price" Nullable(Float64) COMMENT  'quesmaMetadataV1:fieldName=taxful_total_price',
+	"sku" Array(String) COMMENT  'quesmaMetadataV1:fieldName=sku',
+	"taxless_total_price" Nullable(Float64) COMMENT  'quesmaMetadataV1:fieldName=taxless_total_price',
+	"total_unique_products" Nullable(Int64) COMMENT  'quesmaMetadataV1:fieldName=total_unique_products',
+	"geoip_region_name" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=geoip.region_name',
+	"category" Array(String) COMMENT  'quesmaMetadataV1:fieldName=category',
+	"products_created_on" Array(DateTime64) COMMENT  'quesmaMetadataV1:fieldName=products.created_on',
+	"products_taxful_price" Array(Float64) COMMENT  'quesmaMetadataV1:fieldName=products.taxful_price',
+	"user" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=user',
+	"currency" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=currency',
+	"day_of_week" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=day_of_week',
+	"geoip_location_lat" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=geoip.location.lat',
+	"geoip_city_name" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=geoip.city_name',
+	"products__id" Array(String) COMMENT  'quesmaMetadataV1:fieldName=products._id',
+	"products_discount_amount" Array(Int64) COMMENT  'quesmaMetadataV1:fieldName=products.discount_amount',
+	"customer_last_name" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=customer_last_name',
+	"total_quantity" Nullable(Int64) COMMENT  'quesmaMetadataV1:fieldName=total_quantity',
+	"geoip_continent_name" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=geoip.continent_name',
+	"products_price" Array(Float64) COMMENT  'quesmaMetadataV1:fieldName=products.price',
+	"products_sku" Array(String) COMMENT  'quesmaMetadataV1:fieldName=products.sku',
+	"products_discount_percentage" Array(Int64) COMMENT  'quesmaMetadataV1:fieldName=products.discount_percentage',
+	"type" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=type',
+	"customer_full_name" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=customer_full_name',
+	"products_min_price" Array(Float64) COMMENT  'quesmaMetadataV1:fieldName=products.min_price',
+	"products_unit_discount_amount" Array(Int64) COMMENT  'quesmaMetadataV1:fieldName=products.unit_discount_amount',
+	"products_manufacturer" Array(String) COMMENT  'quesmaMetadataV1:fieldName=products.manufacturer',
+	"products_base_unit_price" Array(Float64) COMMENT  'quesmaMetadataV1:fieldName=products.base_unit_price',
+	"day_of_week_i" Nullable(Int64) COMMENT  'quesmaMetadataV1:fieldName=day_of_week_i',
+	"manufacturer" Array(String) COMMENT  'quesmaMetadataV1:fieldName=manufacturer',
+	"customer_first_name" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=customer_first_name',
+	"products_product_id" Array(Int64) COMMENT  'quesmaMetadataV1:fieldName=products.product_id',
+	"customer_gender" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=customer_gender',
+	"email" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=email',
+	"order_id" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=order_id',
+	"customer_phone" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=customer_phone',
+	"order_date" Nullable(DateTime64) COMMENT  'quesmaMetadataV1:fieldName=order_date',
+	"geoip_location_lon" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=geoip.location.lon',
+	"products_base_price" Array(Float64) COMMENT  'quesmaMetadataV1:fieldName=products.base_price',
+	"products_category" Array(String) COMMENT  'quesmaMetadataV1:fieldName=products.category',
+	"products_tax_amount" Array(Int64) COMMENT  'quesmaMetadataV1:fieldName=products.tax_amount',
+	"products_product_name" Array(String) COMMENT  'quesmaMetadataV1:fieldName=products.product_name',
+	"customer_id" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=customer_id',
+	"event_dataset" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=event.dataset',
+	"geoip_country_iso_code" Nullable(String) COMMENT  'quesmaMetadataV1:fieldName=geoip.country_iso_code',
+	"products_taxless_price" Array(Float64) COMMENT  'quesmaMetadataV1:fieldName=products.taxless_price',
+	"products_quantity" Array(Int64) COMMENT  'quesmaMetadataV1:fieldName=products.quantity',
+	"customer_birth_date" Nullable(DateTime64) COMMENT  'quesmaMetadataV1:fieldName=customer_birth_date',
+
+)
+ENGINE = MergeTree
+ORDER BY ("@timestamp")
+
+COMMENT 'created by Quesma'`
+
+	fieldNames := []string{"@timestamp", "attributes_values", "attributes_metadata", "taxful_total_price", "sku", "taxless_total_price", "total_unique_products", "geoip_region_name", "category", "products_created_on", "products_taxful_price", "user", "currency", "day_of_week", "geoip_location_lat", "geoip_city_name", "products__id", "products_discount_amount", "customer_last_name", "total_quantity", "geoip_continent_name", "products_price", "products_sku", "products_discount_percentage", "type", "customer_full_name", "products_min_price", "products_unit_discount_amount", "products_manufacturer", "products_base_unit_price", "day_of_week_i", "manufacturer", "customer_first_name", "products_product_id", "customer_gender", "email", "order_id", "customer_phone", "order_date", "geoip_location_lon", "products_base_price", "products_category", "products_tax_amount", "products_product_name", "customer_id", "event_dataset", "geoip_country_iso_code", "products_taxless_price", "products_quantity", "customer_birth_date"}
+	table, err := clickhouse.NewTable(q, nil)
+	assert.NoError(t, err)
+	assert.Equal(t, len(fieldNames), len(table.Cols))
+	for _, fieldName := range fieldNames {
+		assert.Contains(t, table.Cols, fieldName)
+	}
+	assert.Equal(t, "kibana_sample_data_ecommerce", table.Name)
+	assert.Equal(t, "quesma_cluster", table.ClusterName)
 }
 
 func TestParseLongNestedSchema(t *testing.T) {
