@@ -60,7 +60,14 @@ func main() {
 	// TODO: Experimental feature, move to the configuration after architecture v2
 	const mysql_passthrough_experiment = false
 	if mysql_passthrough_experiment {
-		launchMysqlPassthrough()
+		launchMySqlPassthrough()
+		return
+	}
+
+	// TODO: Experimental feature, move to the configuration after architecture v2
+	const mysql_vitess_experiment = false
+	if mysql_vitess_experiment {
+		launchMySqlVitess()
 		return
 	}
 
@@ -125,7 +132,7 @@ func main() {
 	if cfg.EnableIngest {
 		if cfg.CreateCommonTable {
 			// Ensure common table exists. This table have to be created before ingest processor starts
-			common_table.EnsureCommonTableExists(connectionPool)
+			common_table.EnsureCommonTableExists(connectionPool, cfg.ClusterName)
 		}
 
 		ingestProcessor = ingest.NewIngestProcessor(&cfg, connectionPool, phoneHomeAgent, tableDisco, schemaRegistry, virtualTableStorage, tableResolver)
@@ -159,21 +166,45 @@ func main() {
 
 }
 
-func launchMysqlPassthrough() {
+func launchMySqlVitess() {
+	var frontendConn, err = frontend_connectors.NewVitessMySqlConnector(":13306")
+	if err != nil {
+		panic(err)
+	}
+	var vitessProcessor quesma_api.Processor = processors.NewVitessMySqlProcessor()
+	frontendConn.SetHandlers([]quesma_api.Processor{vitessProcessor})
+	var mySqlBackendConn = backend_connectors.NewMySqlBackendConnector("root:my-secret-pw@tcp(localhost:3306)/exampledb3")
+	var mySqlPipeline quesma_api.PipelineBuilder = quesma_api.NewPipeline()
+	mySqlPipeline.AddProcessor(vitessProcessor)
+	mySqlPipeline.AddFrontendConnector(frontendConn)
+	mySqlPipeline.AddBackendConnector(mySqlBackendConn)
+	var quesmaBuilder quesma_api.QuesmaBuilder = quesma_api.NewQuesma(quesma_api.EmptyDependencies())
+	quesmaBuilder.AddPipeline(mySqlPipeline)
+	qb, err := quesmaBuilder.Build()
+	if err != nil {
+		panic(err)
+	}
+	qb.Start()
+	stop := make(chan os.Signal, 1)
+	<-stop
+	qb.Stop(context.Background())
+}
+
+func launchMySqlPassthrough() {
 	var frontendConn = frontend_connectors.NewTCPConnector(":13306")
-	var tcpProcessor quesma_api.Processor = processors.NewTcpMysqlPassthroughProcessor()
-	var tcpPostgressHandler = frontend_connectors.TcpMysqlConnectionHandler{}
-	frontendConn.AddConnectionHandler(&tcpPostgressHandler)
-	var postgressPipeline quesma_api.PipelineBuilder = quesma_api.NewPipeline()
-	postgressPipeline.AddProcessor(tcpProcessor)
-	postgressPipeline.AddFrontendConnector(frontendConn)
+	var tcpProcessor quesma_api.Processor = processors.NewTcpMySqlPassthroughProcessor()
+	var tcpMySqlHandler = frontend_connectors.TcpMySqlConnectionHandler{}
+	frontendConn.AddConnectionHandler(&tcpMySqlHandler)
+	var mySqlPipeline quesma_api.PipelineBuilder = quesma_api.NewPipeline()
+	mySqlPipeline.AddProcessor(tcpProcessor)
+	mySqlPipeline.AddFrontendConnector(frontendConn)
 	var quesmaBuilder quesma_api.QuesmaBuilder = quesma_api.NewQuesma(quesma_api.EmptyDependencies())
 	backendConn, err := backend_connectors.NewTcpBackendConnector("localhost:3306")
 	if err != nil {
 		panic(err)
 	}
-	postgressPipeline.AddBackendConnector(backendConn)
-	quesmaBuilder.AddPipeline(postgressPipeline)
+	mySqlPipeline.AddBackendConnector(backendConn)
+	quesmaBuilder.AddPipeline(mySqlPipeline)
 	qb, err := quesmaBuilder.Build()
 	if err != nil {
 		panic(err)
