@@ -11,30 +11,30 @@ import (
 	"github.com/QuesmaOrg/quesma/quesma/util"
 	"github.com/goccy/go-json"
 	"reflect"
+	"slices"
 	"strings"
 	"time"
 )
 
-type FieldAtIndex = int // for facets/histogram what Cols[i] means
-
-type QueryResultCol struct {
-	ColName string // quoted, e.g. `"message"`
-	Value   interface{}
-	ColType schema.QuesmaType
-}
+type (
+	QueryResultCol struct {
+		ColName string // quoted, e.g. `"message"`
+		Value   interface{}
+		ColType schema.QuesmaType
+	}
+	QueryResultRow struct {
+		Index string
+		Cols  []QueryResultCol
+	}
+)
 
 func NewQueryResultCol(colName string, value interface{}) QueryResultCol {
 	return QueryResultCol{ColName: colName, Value: value}
 }
 
-type QueryResultRow struct {
-	Index string
-	Cols  []QueryResultCol
-}
-
 // String returns the string representation of the column in format `"<colName>": <value>`, properly quoted.
-func (c QueryResultCol) String(ctx context.Context) string {
-	valueExtracted := c.ExtractValue(ctx)
+func (c *QueryResultCol) String(ctx context.Context) string {
+	valueExtracted := c.ExtractValue()
 	if valueExtracted == nil {
 		return ""
 	}
@@ -62,7 +62,7 @@ func (c QueryResultCol) String(ctx context.Context) string {
 
 // ExtractValue returns the value of the column. If it is a pointer, it returns the value of the pointer.
 // Care: it's untested how it works with '[]type' or '[]*type'.
-func (c QueryResultCol) ExtractValue(ctx context.Context) any {
+func (c *QueryResultCol) ExtractValue() any {
 	if c.Value == nil {
 		return nil
 	}
@@ -76,6 +76,14 @@ func (c QueryResultCol) ExtractValue(ctx context.Context) any {
 	}
 
 	return c.Value
+}
+
+func (c *QueryResultCol) isArray() bool {
+	if c.Value == nil {
+		return false
+	}
+	v := reflect.ValueOf(c.Value)
+	return v.Kind() == reflect.Slice || v.Kind() == reflect.Array
 }
 
 func (r *QueryResultRow) String(ctx context.Context) string {
@@ -110,6 +118,72 @@ func (r *QueryResultRow) Copy() QueryResultRow {
 
 func (r *QueryResultRow) LastColValue() any {
 	return r.Cols[len(r.Cols)-1].Value
+}
+
+// SameSubsetOfColumns returns if r and other have the same values for columns with names in colNames
+// They are results of the same query, so we can assume that the columns are in the same order.
+func (r *QueryResultRow) SameSubsetOfColumns(other *QueryResultRow, colNames []string) bool {
+	for i := range min(len(r.Cols), len(other.Cols)) {
+		if slices.Contains(colNames, r.Cols[i].ColName) {
+			isArray1 := r.Cols[i].isArray()
+			isArray2 := other.Cols[i].isArray()
+
+			if !isArray1 && !isArray2 {
+				if r.Cols[i].ExtractValue() != other.Cols[i].ExtractValue() {
+					return false
+				}
+			} else if isArray1 && isArray2 {
+				if !reflect.DeepEqual(r.Cols[i].Value, other.Cols[i].Value) {
+					return false
+				}
+			} else {
+				return false
+			}
+		}
+	}
+	return true
+}
+
+// firstNColumnsHaveSameValues returns if 2 rows have the same values for the first N columns
+func (r *QueryResultRow) firstNColumnsHaveSameValues(other *QueryResultRow, N int) bool {
+	for i := 0; i < N; i++ {
+		isArray1 := r.Cols[i].isArray()
+		isArray2 := other.Cols[i].isArray()
+
+		if !isArray1 && !isArray2 {
+			if r.Cols[i].ExtractValue() != other.Cols[i].ExtractValue() {
+				return false
+			}
+		} else if isArray1 && isArray2 {
+			if !reflect.DeepEqual(r.Cols[i].Value, other.Cols[i].Value) {
+				return false
+			}
+		} else {
+			return false
+		}
+	}
+	return true
+}
+
+// SplitResultSetIntoBuckets splits ResultSet into buckets, based on the first N + 1 columns
+// E.g. if N == 0, we split into buckets based on the first field,
+// e.g. [row(1, ...), row(1, ...), row(2, ...), row(2, ...), row(3, ...)] -> [[row(1, ...), row(1, ...)], [row(2, ...), row(2, ...)], [row(3, ...)]]
+func SplitResultSetIntoBuckets(ResultSet []QueryResultRow, N int) [][]QueryResultRow {
+	if len(ResultSet) == 0 {
+		return [][]QueryResultRow{{}}
+	}
+
+	lastRow := ResultSet[0]
+	buckets := [][]QueryResultRow{{lastRow}}
+	for _, row := range ResultSet[1:] {
+		if row.firstNColumnsHaveSameValues(&lastRow, N) {
+			buckets[len(buckets)-1] = append(buckets[len(buckets)-1], row)
+		} else {
+			buckets = append(buckets, []QueryResultRow{row})
+		}
+		lastRow = row
+	}
+	return buckets
 }
 
 func FirstNonNilIndex(rows []QueryResultRow) int {
