@@ -82,6 +82,7 @@ type BackendConnector struct {
 	Config RelationalDbConfiguration `koanf:"config"`
 }
 
+// RelationalDbConfiguration works fine for non-relational databases too, consider rename
 type RelationalDbConfiguration struct {
 	//ConnectorName string `koanf:"name"`
 	ConnectorType string `koanf:"type"`
@@ -89,6 +90,7 @@ type RelationalDbConfiguration struct {
 	User          string `koanf:"user"`
 	Password      string `koanf:"password"`
 	Database      string `koanf:"database"`
+	ClusterName   string `koanf:"clusterName"` // When creating tables by Quesma - they'll use `ON CLUSTER ClusterName` clause
 	AdminUrl      *Url   `koanf:"adminUrl"`
 	DisableTLS    bool   `koanf:"disableTLS"`
 }
@@ -115,6 +117,8 @@ const DefaultWildcardIndexName = "*"
 type QuesmaProcessorConfig struct {
 	UseCommonTable bool           `koanf:"useCommonTable"`
 	IndexConfig    IndicesConfigs `koanf:"indexes"`
+	// DefaultTargetConnectorType is used in V2 code only
+	DefaultTargetConnectorType string //it is not serialized to maintain configuration BWC, so it's basically just populated from '*' config in `config_v2.go`
 }
 
 type IndicesConfigs map[string]IndexConfiguration
@@ -163,9 +167,27 @@ func (c *QuesmaNewConfiguration) Validate() error {
 	return nil
 }
 
-func (c *QuesmaNewConfiguration) getFrontendConnectorByName(name string) *FrontendConnector {
+func (c *QuesmaNewConfiguration) GetFrontendConnectorByName(name string) *FrontendConnector {
 	for _, fc := range c.FrontendConnectors {
 		if fc.Name == name {
+			return &fc
+		}
+	}
+	return nil
+}
+
+func (c *QuesmaNewConfiguration) GetFrontendConnectorByType(typ string) *FrontendConnector {
+	for _, fc := range c.FrontendConnectors {
+		if fc.Type == typ {
+			return &fc
+		}
+	}
+	return nil
+}
+
+func (c *QuesmaNewConfiguration) GetBackendConnectorByType(typ string) *BackendConnector {
+	for _, fc := range c.BackendConnectors {
+		if fc.Type == typ {
 			return &fc
 		}
 	}
@@ -211,11 +233,11 @@ func (c *QuesmaNewConfiguration) validatePipelines() error {
 		// We plan to only support a case of a single pipeline for querying (this code validates this).
 		// However, we haven't yet implemented the case of disabling ingest, so single pipeline case is not yet supported.
 		fcName := c.Pipelines[0].FrontendConnectors[0]
-		if fc := c.getFrontendConnectorByName(fcName); fc != nil {
+		if fc := c.GetFrontendConnectorByName(fcName); fc != nil {
 			if fc.Type != ElasticsearchFrontendQueryConnectorName {
 				return fmt.Errorf("single-pipeline Quesma can only be used for querying, but the frontend connector is not of query type")
 			}
-			proc := c.getProcessorByName(c.Pipelines[0].Processors[0])
+			proc := c.GetProcessorByName(c.Pipelines[0].Processors[0])
 			if proc == nil {
 				return fmt.Errorf("processor named [%s] not found in configuration", c.Pipelines[0].Processors[0])
 			}
@@ -224,7 +246,7 @@ func (c *QuesmaNewConfiguration) validatePipelines() error {
 				if len(declaredBackendConnectors) != 1 {
 					return fmt.Errorf("noop processor supports only one backend connector")
 				}
-				if conn := c.getBackendConnectorByName(declaredBackendConnectors[0]); conn.Type != ElasticsearchBackendConnectorName {
+				if conn := c.GetBackendConnectorByName(declaredBackendConnectors[0]); conn.Type != ElasticsearchBackendConnectorName {
 					return fmt.Errorf("noop processor can be connected only to elasticsearch backend connector")
 				}
 			} else if proc.Type == QuesmaV1ProcessorQuery {
@@ -233,7 +255,7 @@ func (c *QuesmaNewConfiguration) validatePipelines() error {
 				}
 				var backendConnectorTypes []string
 				for _, con := range declaredBackendConnectors {
-					connector := c.getBackendConnectorByName(con)
+					connector := c.GetBackendConnectorByName(con)
 					if connector == nil {
 						return fmt.Errorf("backend connector named [%s] not found in configuration", con)
 					}
@@ -259,7 +281,7 @@ func (c *QuesmaNewConfiguration) validatePipelines() error {
 		}
 	}
 	if isDualPipeline {
-		fc1, fc2 := c.getFrontendConnectorByName(c.Pipelines[0].FrontendConnectors[0]), c.getFrontendConnectorByName(c.Pipelines[1].FrontendConnectors[0])
+		fc1, fc2 := c.GetFrontendConnectorByName(c.Pipelines[0].FrontendConnectors[0]), c.GetFrontendConnectorByName(c.Pipelines[1].FrontendConnectors[0])
 		if fc1 == nil {
 			return fmt.Errorf("frontend connector named [%s] not found in configuration", c.Pipelines[0].FrontendConnectors[0])
 		}
@@ -276,14 +298,14 @@ func (c *QuesmaNewConfiguration) validatePipelines() error {
 		} else {
 			queryPipeline, ingestPipeline = c.Pipelines[1], c.Pipelines[0]
 		}
-		ingestProcessor := c.getProcessorByName(ingestPipeline.Processors[0])
+		ingestProcessor := c.GetProcessorByName(ingestPipeline.Processors[0])
 		if ingestProcessor == nil {
 			return fmt.Errorf("ingest processor named [%s] not found in configuration", ingestPipeline.Processors[0])
 		}
 		if ingestProcessor.Type != QuesmaV1ProcessorIngest && ingestProcessor.Type != QuesmaV1ProcessorNoOp {
 			return fmt.Errorf("ingest pipeline must have ingest-type or noop processor")
 		}
-		queryProcessor := c.getProcessorByName(queryPipeline.Processors[0])
+		queryProcessor := c.GetProcessorByName(queryPipeline.Processors[0])
 		if queryProcessor == nil {
 			return fmt.Errorf("query processor named [%s] not found in configuration", ingestPipeline.Processors[0])
 		}
@@ -396,7 +418,7 @@ func (c *QuesmaNewConfiguration) validateProcessor(p Processor) error {
 				return errTarget
 			}
 			for _, target := range targets {
-				if c.getBackendConnectorByName(target.target) == nil {
+				if c.GetBackendConnectorByName(target.target) == nil {
 					return fmt.Errorf("invalid target %s in configuration of index %s", target, indexName)
 				}
 			}
@@ -437,19 +459,19 @@ func (c *QuesmaNewConfiguration) validatePipeline(pipeline Pipeline) error {
 	if !slices.Contains(c.definedProcessorNames(), pipeline.Processors[0]) {
 		errAcc = multierror.Append(errAcc, fmt.Errorf("processor named %s referenced in %s not found in configuration", pipeline.Processors[0], pipeline.Name))
 	} else {
-		onlyProcessorInPipeline := c.getProcessorByName(pipeline.Processors[0])
+		onlyProcessorInPipeline := c.GetProcessorByName(pipeline.Processors[0])
 		if onlyProcessorInPipeline.Type == QuesmaV1ProcessorNoOp {
 			if len(pipeline.BackendConnectors) != 1 {
 				return multierror.Append(errAcc, fmt.Errorf("pipeline %s has a noop processor supports only one backend connector", pipeline.Name))
 			}
-			if conn := c.getBackendConnectorByName(pipeline.BackendConnectors[0]); conn.Type != ElasticsearchBackendConnectorName {
+			if conn := c.GetBackendConnectorByName(pipeline.BackendConnectors[0]); conn.Type != ElasticsearchBackendConnectorName {
 				return multierror.Append(errAcc, fmt.Errorf("pipeline %s has a noop processor which can be connected only to elasticsearch backend connector", pipeline.Name))
 			}
 		}
 		if onlyProcessorInPipeline.Type == QuesmaV1ProcessorQuery || onlyProcessorInPipeline.Type == QuesmaV1ProcessorIngest {
 			foundElasticBackendConnector := false
 			for _, backendConnectorName := range pipeline.BackendConnectors {
-				backendConnector := c.getBackendConnectorByName(backendConnectorName)
+				backendConnector := c.GetBackendConnectorByName(backendConnectorName)
 				if backendConnector == nil {
 					return multierror.Append(errAcc, fmt.Errorf("backend connector named %s referenced in %s not found in configuration", backendConnectorName, pipeline.Name))
 				}
@@ -466,7 +488,7 @@ func (c *QuesmaNewConfiguration) validatePipeline(pipeline Pipeline) error {
 	return errAcc
 }
 
-func (c *QuesmaNewConfiguration) getBackendConnectorByName(name string) *BackendConnector {
+func (c *QuesmaNewConfiguration) GetBackendConnectorByName(name string) *BackendConnector {
 	for _, b := range c.BackendConnectors {
 		if b.Name == name {
 			return &b
@@ -475,7 +497,7 @@ func (c *QuesmaNewConfiguration) getBackendConnectorByName(name string) *Backend
 	return nil
 }
 
-func (c *QuesmaNewConfiguration) getProcessorByName(name string) *Processor {
+func (c *QuesmaNewConfiguration) GetProcessorByName(name string) *Processor {
 	for _, p := range c.Processors {
 		if p.Name == name {
 			return &p
@@ -484,8 +506,17 @@ func (c *QuesmaNewConfiguration) getProcessorByName(name string) *Processor {
 	return nil
 }
 
+func (c *QuesmaNewConfiguration) GetProcessorByType(typ ProcessorType) *Processor {
+	for _, p := range c.Processors {
+		if p.Type == typ {
+			return &p
+		}
+	}
+	return nil
+}
+
 func (c *QuesmaNewConfiguration) getTargetType(backendConnectorName string) (string, bool) {
-	backendConnector := c.getBackendConnectorByName(backendConnectorName)
+	backendConnector := c.GetBackendConnectorByName(backendConnectorName)
 	if backendConnector == nil {
 		return "", false
 	}
@@ -539,7 +570,7 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 	isSinglePipeline, isDualPipeline := c.getPipelinesType()
 
 	if isSinglePipeline {
-		processor := c.getProcessorByName(c.Pipelines[0].Processors[0])
+		processor := c.GetProcessorByName(c.Pipelines[0].Processors[0])
 		procType := processor.Type
 		if procType == QuesmaV1ProcessorNoOp {
 			conf.TransparentProxy = true
@@ -560,7 +591,9 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 			}
 			for _, target := range targets {
 				if targetType, found := c.getTargetType(target.target); found {
-					defaultConfig.QueryTarget = append(defaultConfig.QueryTarget, targetType)
+					if !slices.Contains(defaultConfig.QueryTarget, targetType) {
+						defaultConfig.QueryTarget = append(defaultConfig.QueryTarget, targetType)
+					}
 				} else {
 					errAcc = multierror.Append(errAcc, fmt.Errorf("invalid target %s in configuration of %s", target, DefaultWildcardIndexName))
 				}
@@ -589,6 +622,43 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 			conf.DefaultIngestTarget = []string{}
 			conf.DefaultQueryTarget = defaultConfig.QueryTarget
 			conf.AutodiscoveryEnabled = slices.Contains(conf.DefaultQueryTarget, ClickhouseTarget)
+
+			// safe to call per validation earlier
+			if targts, ok := queryProcessor.Config.IndexConfig[DefaultWildcardIndexName].Target.([]interface{}); ok {
+				conn := c.GetBackendConnectorByName(targts[0].(string))
+				queryProcessor.Config.DefaultTargetConnectorType = conn.Type
+				for indexName, indexCfg := range queryProcessor.Config.IndexConfig {
+					var targetType string
+					indexTarget, ok2 := indexCfg.Target.([]interface{})
+					if len(indexTarget) > 0 && ok2 {
+						tgt := indexTarget[0]
+						var indexTargetType *BackendConnector
+						if targetMap, ok := tgt.(map[string]interface{}); ok {
+							for name := range targetMap {
+								indexTargetType = c.GetBackendConnectorByName(name)
+								break
+							}
+						} else {
+							indexTargetType = c.GetBackendConnectorByName(indexTarget[0].(string))
+						}
+
+						if indexTargetType.Type == "elasticsearch" {
+							targetType = "elasticsearch"
+						} else { // clickhouse-os, hydrolix included
+							targetType = "clickhouse"
+						}
+						indexCfg.QueryTarget = []string{targetType}
+						queryProcessor.Config.IndexConfig[indexName] = indexCfg
+					}
+				}
+				queryProcessor.Config.DefaultTargetConnectorType = conn.Type
+				c.updateProcessorConfig(queryProcessor.Name, queryProcessor.Config)
+			}
+			if defaultQueryConfig, ok := queryProcessor.Config.IndexConfig[DefaultWildcardIndexName]; ok {
+				conf.DefaultQueryOptimizers = defaultQueryConfig.Optimizers
+			} else {
+				conf.DefaultQueryOptimizers = nil
+			}
 			delete(queryProcessor.Config.IndexConfig, DefaultWildcardIndexName)
 
 			for indexName, indexConfig := range queryProcessor.Config.IndexConfig {
@@ -599,7 +669,9 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 				}
 				for _, target := range targets {
 					if targetType, found := c.getTargetType(target.target); found {
-						processedConfig.QueryTarget = append(processedConfig.QueryTarget, targetType)
+						if !slices.Contains(processedConfig.QueryTarget, targetType) {
+							processedConfig.QueryTarget = append(processedConfig.QueryTarget, targetType)
+						}
 					} else {
 						errAcc = multierror.Append(errAcc, fmt.Errorf("invalid target %s in configuration of index %s", target, indexName))
 					}
@@ -639,14 +711,14 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 	}
 
 	if isDualPipeline {
-		fc1 := c.getFrontendConnectorByName(c.Pipelines[0].FrontendConnectors[0])
+		fc1 := c.GetFrontendConnectorByName(c.Pipelines[0].FrontendConnectors[0])
 		var queryPipeline, ingestPipeline Pipeline
 		if fc1.Type == ElasticsearchFrontendQueryConnectorName {
 			queryPipeline, ingestPipeline = c.Pipelines[0], c.Pipelines[1]
 		} else {
 			queryPipeline, ingestPipeline = c.Pipelines[1], c.Pipelines[0]
 		}
-		queryProcessor, ingestProcessor := c.getProcessorByName(queryPipeline.Processors[0]), c.getProcessorByName(ingestPipeline.Processors[0])
+		queryProcessor, ingestProcessor := c.GetProcessorByName(queryPipeline.Processors[0]), c.GetProcessorByName(ingestPipeline.Processors[0])
 
 		if queryProcessor.Type == QuesmaV1ProcessorNoOp && ingestProcessor.Type == QuesmaV1ProcessorNoOp {
 			conf.TransparentProxy = true
@@ -663,7 +735,9 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 		}
 		for _, target := range targets {
 			if targetType, found := c.getTargetType(target.target); found {
-				defaultConfig.QueryTarget = append(defaultConfig.QueryTarget, targetType)
+				if !slices.Contains(defaultConfig.QueryTarget, targetType) {
+					defaultConfig.QueryTarget = append(defaultConfig.QueryTarget, targetType)
+				}
 			} else {
 				errAcc = multierror.Append(errAcc, fmt.Errorf("invalid target %s in configuration of %s", target, DefaultWildcardIndexName))
 			}
@@ -693,7 +767,9 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 		}
 		for _, target := range targets {
 			if targetType, found := c.getTargetType(target.target); found {
-				defaultConfig.IngestTarget = append(defaultConfig.IngestTarget, targetType)
+				if !slices.Contains(defaultConfig.IngestTarget, targetType) {
+					defaultConfig.IngestTarget = append(defaultConfig.IngestTarget, targetType)
+				}
 			} else {
 				errAcc = multierror.Append(errAcc, fmt.Errorf("invalid target %s in configuration of %s", target, DefaultWildcardIndexName))
 			}
@@ -729,6 +805,46 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 		conf.DefaultQueryTarget = defaultConfig.QueryTarget
 		conf.AutodiscoveryEnabled = slices.Contains(conf.DefaultQueryTarget, ClickhouseTarget)
 
+		// we're calling this here because we don't allow having one ingest-only pipeline.
+		if relationalDBErr == nil && relDBConn != nil {
+			conf.ClusterName = relDBConn.ClusterName
+		}
+
+		// safe to call per validation earlier
+		if targts, ok := queryProcessor.Config.IndexConfig[DefaultWildcardIndexName].Target.([]interface{}); ok {
+			conn := c.GetBackendConnectorByName(targts[0].(string))
+			queryProcessor.Config.DefaultTargetConnectorType = conn.Type
+			for indexName, indexCfg := range queryProcessor.Config.IndexConfig {
+				var targetType string
+				indexTarget, ok2 := indexCfg.Target.([]interface{})
+				if len(indexTarget) > 0 && ok2 {
+					tgt := indexTarget[0]
+					var indexTargetType *BackendConnector
+					if targetMap, ok := tgt.(map[string]interface{}); ok {
+						for name := range targetMap {
+							indexTargetType = c.GetBackendConnectorByName(name)
+							break
+						}
+					} else {
+						indexTargetType = c.GetBackendConnectorByName(indexTarget[0].(string))
+					}
+
+					if indexTargetType.Type == "elasticsearch" {
+						targetType = "elasticsearch"
+					} else { // clickhouse-os, hydrolix included
+						targetType = "clickhouse"
+					}
+					indexCfg.QueryTarget = []string{targetType}
+					queryProcessor.Config.IndexConfig[indexName] = indexCfg
+				}
+			}
+			c.updateProcessorConfig(queryProcessor.Name, queryProcessor.Config)
+		}
+		if defaultQueryConfig, ok := queryProcessor.Config.IndexConfig[DefaultWildcardIndexName]; ok {
+			conf.DefaultQueryOptimizers = defaultQueryConfig.Optimizers
+		} else {
+			conf.DefaultQueryOptimizers = nil
+		}
 		delete(queryProcessor.Config.IndexConfig, DefaultWildcardIndexName)
 
 		for indexName, indexConfig := range queryProcessor.Config.IndexConfig {
@@ -741,7 +857,9 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 			}
 			for _, target := range targets {
 				if targetType, found := c.getTargetType(target.target); found {
-					processedConfig.QueryTarget = append(processedConfig.QueryTarget, targetType)
+					if !slices.Contains(processedConfig.QueryTarget, targetType) {
+						processedConfig.QueryTarget = append(processedConfig.QueryTarget, targetType)
+					}
 				} else {
 					errAcc = multierror.Append(errAcc, fmt.Errorf("invalid target %s in configuration of index %s", target, indexName))
 				}
@@ -784,6 +902,37 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 			conf.DefaultIngestOptimizers = nil
 		}
 
+		// safe to call per validation earlier
+		if targts, ok := ingestProcessor.Config.IndexConfig[DefaultWildcardIndexName].Target.([]interface{}); ok {
+			conn := c.GetBackendConnectorByName(targts[0].(string))
+			ingestProcessor.Config.DefaultTargetConnectorType = conn.Type
+			// In order to maintain compat with v1 code we have to fill QueryTarget and IngestTarget for each index
+			for indexName, indexCfg := range ingestProcessor.Config.IndexConfig {
+				var targetType string
+				indexTarget, ok2 := indexCfg.Target.([]interface{})
+				if len(indexTarget) > 0 && ok2 {
+					tgt := indexTarget[0]
+					var indexTargetType *BackendConnector
+					if targetMap, ok := tgt.(map[string]interface{}); ok {
+						for name := range targetMap {
+							indexTargetType = c.GetBackendConnectorByName(name)
+							break
+						}
+					} else {
+						indexTargetType = c.GetBackendConnectorByName(indexTarget[0].(string))
+					}
+
+					if indexTargetType.Type == "elasticsearch" {
+						targetType = "elasticsearch"
+					} else { // clickhouse-os, hydrolix included
+						targetType = "clickhouse"
+					}
+					indexCfg.IngestTarget = []string{targetType}
+					ingestProcessor.Config.IndexConfig[indexName] = indexCfg
+				}
+			}
+			c.updateProcessorConfig(ingestProcessor.Name, ingestProcessor.Config)
+		}
 		delete(ingestProcessor.Config.IndexConfig, DefaultWildcardIndexName)
 
 		for indexName, indexConfig := range ingestProcessor.Config.IndexConfig {
@@ -802,7 +951,9 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 			}
 			for _, target := range targets {
 				if targetType, found := c.getTargetType(target.target); found {
-					processedConfig.IngestTarget = append(processedConfig.IngestTarget, targetType)
+					if !slices.Contains(processedConfig.IngestTarget, targetType) {
+						processedConfig.IngestTarget = append(processedConfig.IngestTarget, targetType)
+					}
 				} else {
 					errAcc = multierror.Append(errAcc, fmt.Errorf("invalid target %s in configuration of index %s", target, indexName))
 				}
@@ -833,7 +984,6 @@ func (c *QuesmaNewConfiguration) TranslateToLegacyConfig() QuesmaConfiguration {
 	}
 
 END:
-
 	for _, idxCfg := range conf.IndexConfig {
 		if idxCfg.UseCommonTable {
 			conf.CreateCommonTable = true
@@ -968,4 +1118,14 @@ func (c *QuesmaNewConfiguration) getTargetsExtendedConfig(target any) ([]struct 
 		}
 	}
 	return result, nil
+}
+
+func (c *QuesmaNewConfiguration) updateProcessorConfig(processorName string, newConfig QuesmaProcessorConfig) error {
+	for i, p := range c.Processors {
+		if p.Name == processorName {
+			c.Processors[i].Config = newConfig
+			return nil
+		}
+	}
+	return fmt.Errorf("processor named %s not found in configuration", processorName)
 }
