@@ -15,7 +15,7 @@ import (
 	"github.com/QuesmaOrg/quesma/quesma/model"
 	"github.com/QuesmaOrg/quesma/quesma/optimize"
 	"github.com/QuesmaOrg/quesma/quesma/painful"
-	"github.com/QuesmaOrg/quesma/quesma/queryparser"
+	"github.com/QuesmaOrg/quesma/quesma/parsers/elastic_query_dsl"
 	"github.com/QuesmaOrg/quesma/quesma/quesma/async_search_storage"
 	"github.com/QuesmaOrg/quesma/quesma/quesma/config"
 	"github.com/QuesmaOrg/quesma/quesma/quesma/errors"
@@ -57,7 +57,7 @@ type QueryRunner struct {
 	// this is passed to the QueryTranslator to render date math expressions
 	DateMathRenderer         string // "clickhouse_interval" or "literal"  if not set, we use "clickhouse_interval"
 	currentParallelQueryJobs atomic.Int64
-	transformationPipeline   TransformationPipeline
+	transformationPipeline   model.TransformationPipeline
 	schemaRegistry           schema.Registry
 	ABResultsSender          ab_testing.Sender
 	tableResolver            table_resolver.TableResolver
@@ -77,10 +77,11 @@ type QueryRunnerIFace interface {
 	GetLogManager() clickhouse.LogManagerIFace
 	DeleteAsyncSearch(id string) ([]byte, error)
 	HandlePartialAsyncSearch(ctx context.Context, id string) ([]byte, error)
+	HandleMultiSearch(ctx context.Context, defaultIndexName string, body types.NDJSON) ([]byte, error)
 }
 
 func (q *QueryRunner) EnableQueryOptimization(cfg *config.QuesmaConfiguration) {
-	q.transformationPipeline.transformers = append(q.transformationPipeline.transformers, optimize.NewOptimizePipeline(cfg))
+	q.transformationPipeline.AddTransformer(optimize.NewOptimizePipeline(cfg))
 }
 
 func NewQueryRunner(lm clickhouse.LogManagerIFace,
@@ -92,19 +93,18 @@ func NewQueryRunner(lm clickhouse.LogManagerIFace,
 	tableDiscovery clickhouse.TableDiscovery) *QueryRunner {
 
 	ctx, cancel := context.WithCancel(context.Background())
-
+	transformationPipeline := model.NewTransformationPipeline()
+	transformationPipeline.AddTransformer(NewSchemaCheckPass(cfg, tableDiscovery, defaultSearchAfterStrategy))
 	return &QueryRunner{logManager: lm, cfg: cfg, debugInfoCollector: qmc,
 		executionCtx: ctx, cancel: cancel,
-		AsyncRequestStorage:  async_search_storage.NewAsyncSearchStorageInMemory(),
-		AsyncQueriesContexts: async_search_storage.NewAsyncQueryContextStorageInMemory(),
-		transformationPipeline: TransformationPipeline{
-			transformers: []model.QueryTransformer{NewSchemaCheckPass(cfg, tableDiscovery, defaultSearchAfterStrategy)},
-		},
-		schemaRegistry:     schemaRegistry,
-		ABResultsSender:    abResultsRepository,
-		tableResolver:      resolver,
-		tableDiscovery:     tableDiscovery,
-		maxParallelQueries: maxParallelQueries,
+		AsyncRequestStorage:    async_search_storage.NewAsyncSearchStorageInMemory(),
+		AsyncQueriesContexts:   async_search_storage.NewAsyncQueryContextStorageInMemory(),
+		transformationPipeline: *transformationPipeline,
+		schemaRegistry:         schemaRegistry,
+		ABResultsSender:        abResultsRepository,
+		tableResolver:          resolver,
+		tableDiscovery:         tableDiscovery,
+		maxParallelQueries:     maxParallelQueries,
 	}
 }
 
@@ -157,6 +157,7 @@ func (q *QueryRunner) HandleCount(ctx context.Context, indexPattern string) (int
 		}
 	}
 
+	// Query execution block
 	if len(indexes) == 1 {
 		return q.logManager.Count(ctx, indexes[0])
 	} else {
@@ -164,13 +165,12 @@ func (q *QueryRunner) HandleCount(ctx context.Context, indexPattern string) (int
 	}
 }
 
-func (q *QueryRunner) HandleMultiSearch(ctx context.Context, defaultIndexName string, body types.NDJSON) ([]byte, error) {
+type msearchQuery struct {
+	indexName string
+	query     types.JSON
+}
 
-	type msearchQuery struct {
-		indexName string
-		query     types.JSON
-	}
-
+func SplitQueries(body types.NDJSON, defaultIndexName string) ([]msearchQuery, error) {
 	var queries []msearchQuery
 
 	var currentQuery *msearchQuery
@@ -211,14 +211,37 @@ func (q *QueryRunner) HandleMultiSearch(ctx context.Context, defaultIndexName st
 		currentQuery = nil
 
 	}
+	return queries, nil
+}
 
+// Composite search is a search that can contain multiple queries
+// like bulk
+func (q *QueryRunner) HandleMultiSearch(ctx context.Context, defaultIndexName string, body types.NDJSON) ([]byte, error) {
+
+	// Split the body into queries
+	// This is what ParseQuery does or should do
+	queries, err := SplitQueries(body, defaultIndexName)
+
+	if err != nil {
+		return nil, err
+	}
+
+	// Handle each query
 	var responses []any
 
+	var queriedIndices []string
 	for _, query := range queries {
-
-		// TODO ask table resolver here and go to the right connector or connectors
-
-		responseBody, err := q.HandleSearch(ctx, query.indexName, query.query)
+		queriedIndices = append(queriedIndices, query.indexName)
+	}
+	logger.DebugWithCtx(ctx).Msgf("handling multisearch: queries=%d, indices=[%s], defaultIndex=[%s]", len(queries), queriedIndices, defaultIndexName)
+	for _, query := range queries {
+		var responseBody []byte
+		if q.shouldRouteQueryToElasticsearch(query) { // this branch is here to get response from multi-search query targeted an index not stored in Clickhouse
+			// this is also a shortcut that we took to delay a bigger refactor, eventually HandleMultiSearch should dispatch all individual queries to proper connector, similarly to `_bulk` endpoint
+			responseBody, err = q.forwardToElasticsearch(ctx, query.indexName, query.query)
+		} else {
+			responseBody, err = q.HandleSearch(ctx, query.indexName, query.query)
+		}
 
 		if err != nil {
 
@@ -230,16 +253,16 @@ func (q *QueryRunner) HandleMultiSearch(ctx context.Context, defaultIndexName st
 				wrappedErr = &quesma_api.Result{StatusCode: http.StatusNotFound}
 			} else if errors.Is(err, quesma_errors.ErrCouldNotParseRequest()) {
 				wrappedErr = &quesma_api.Result{
-					Body:          string(queryparser.BadRequestParseError(err)),
+					Body:          string(elastic_query_dsl.BadRequestParseError(err)),
 					StatusCode:    http.StatusBadRequest,
-					GenericResult: queryparser.BadRequestParseError(err),
+					GenericResult: elastic_query_dsl.BadRequestParseError(err),
 				}
 			} else {
 				logger.ErrorWithCtx(ctx).Msgf("error handling multisearch: %v", err)
 				wrappedErr = &quesma_api.Result{
 					Body:          "Internal error",
 					StatusCode:    http.StatusInternalServerError,
-					GenericResult: queryparser.BadRequestParseError(err),
+					GenericResult: elastic_query_dsl.BadRequestParseError(err),
 				}
 			}
 
@@ -255,6 +278,7 @@ func (q *QueryRunner) HandleMultiSearch(ctx context.Context, defaultIndexName st
 
 	}
 
+	// build the response
 	type msearchResponse struct {
 		Responses []any `json:"responses"`
 	}
@@ -269,12 +293,34 @@ func (q *QueryRunner) HandleMultiSearch(ctx context.Context, defaultIndexName st
 	return responseBody, nil
 }
 
+func (q *QueryRunner) forwardToElasticsearch(ctx context.Context, indexName string, query types.JSON) ([]byte, error) {
+	logger.DebugWithCtx(ctx).Msgf("_msearch query to index=%s forwarded to Elasticsearch", indexName)
+	esClient := elasticsearch.NewSimpleClient(&q.cfg.Elasticsearch)
+	queryBody, _ := query.Bytes()
+	if resp, err := esClient.Request(ctx, "POST", "/_search", queryBody); err != nil {
+		return []byte{}, err
+	} else {
+		respBody, errRead := util.ReadResponseBody(resp)
+		return respBody, errRead
+	}
+}
+
+func (q *QueryRunner) shouldRouteQueryToElasticsearch(query msearchQuery) bool {
+	decision := q.tableResolver.Resolve(quesma_api.QueryPipeline, query.indexName)
+	if len(decision.UseConnectors) == 1 {
+		_, useElastic := decision.UseConnectors[0].(*quesma_api.ConnectorDecisionElastic)
+		return useElastic
+	}
+	return false
+}
+
 func (q *QueryRunner) HandleSearch(ctx context.Context, indexPattern string, body types.JSON) ([]byte, error) {
 	return q.handleSearchCommon(ctx, indexPattern, body, nil)
 }
 
 func (q *QueryRunner) HandleAsyncSearch(ctx context.Context, indexPattern string, body types.JSON,
 	waitForResultsMs int, keepOnCompletion bool) ([]byte, error) {
+	// AsyncQuery marker should be result of ParseQuery
 	async := AsyncQuery{
 		asyncId:          tracing.GetAsyncId(),
 		waitForResultsMs: waitForResultsMs,
@@ -299,7 +345,7 @@ type AsyncQuery struct {
 	startTime        time.Time
 }
 
-func (q *QueryRunner) transformQueries(ctx context.Context, plan *model.ExecutionPlan) error {
+func (q *QueryRunner) transformQueries(plan *model.ExecutionPlan) error {
 	var err error
 	plan.Queries, err = q.transformationPipeline.Transform(plan.Queries)
 	if err != nil {
@@ -358,11 +404,6 @@ func (q *QueryRunner) executePlan(ctx context.Context, plan *model.ExecutionPlan
 		}
 	}
 
-	err = q.transformQueries(ctx, plan)
-	if err != nil {
-		return responseBody, err
-	}
-
 	q.runExecutePlanAsync(ctx, plan, queryTranslator, table, doneCh, optAsync)
 
 	if optAsync == nil {
@@ -408,18 +449,18 @@ func (q *QueryRunner) handleSearchCommon(ctx context.Context, indexPattern strin
 
 		var resp []byte
 		if optAsync != nil {
-			resp, _ = queryparser.EmptyAsyncSearchResponse(optAsync.asyncId, false, 200)
+			resp, _ = elastic_query_dsl.EmptyAsyncSearchResponse(optAsync.asyncId, false, 200)
 		} else {
-			resp = queryparser.EmptySearchResponse(ctx)
+			resp = elastic_query_dsl.EmptySearchResponse(ctx)
 		}
 		return resp, decision.Err
 	}
 
 	if decision.IsEmpty {
 		if optAsync != nil {
-			return queryparser.EmptyAsyncSearchResponse(optAsync.asyncId, false, 200)
+			return elastic_query_dsl.EmptyAsyncSearchResponse(optAsync.asyncId, false, 200)
 		} else {
-			return queryparser.EmptySearchResponse(ctx), nil
+			return elastic_query_dsl.EmptySearchResponse(ctx), nil
 		}
 	}
 
@@ -440,16 +481,18 @@ func (q *QueryRunner) handleSearchCommon(ctx context.Context, indexPattern strin
 			clickhouseConnector = c
 
 		case *quesma_api.ConnectorDecisionElastic:
-			// NOP
+			// After https://github.com/QuesmaOrg/quesma/pull/1278 we should never land in this situation,
+			// previously this was an escape hatch for `_msearch` payload containing Elasticsearch-targetted query
+			// This code lives only to postpone bigger refactor of `handleSearchCommon` which also supports async and A/B testing
 
 		default:
 			return nil, fmt.Errorf("unknown connector type: %T", c)
 		}
 	}
 
-	// it's impossible here to don't have a clickhouse decision
 	if clickhouseConnector == nil {
-		return nil, fmt.Errorf("no clickhouse connector")
+		logger.Warn().Msgf("multi-search payload contains Elasticsearch-targetted query")
+		return nil, fmt.Errorf("quesma-processed _msearch payload contains Elasticsearch-targetted query")
 	}
 
 	var responseBody []byte
@@ -511,9 +554,9 @@ func (q *QueryRunner) handleSearchCommon(ctx context.Context, indexPattern strin
 
 		if len(resolvedIndexes) == 0 {
 			if optAsync != nil {
-				return queryparser.EmptyAsyncSearchResponse(optAsync.asyncId, false, 200)
+				return elastic_query_dsl.EmptyAsyncSearchResponse(optAsync.asyncId, false, 200)
 			} else {
-				return queryparser.EmptySearchResponse(ctx), nil
+				return elastic_query_dsl.EmptySearchResponse(ctx), nil
 			}
 		}
 
@@ -567,7 +610,10 @@ func (q *QueryRunner) handleSearchCommon(ctx context.Context, indexPattern strin
 		pushSecondaryInfo(q.debugInfoCollector, id, "", path, bodyAsBytes, queriesBody, responseBody, startTime)
 		return responseBody, errors.New(string(responseBody))
 	}
-
+	err = q.transformQueries(plan)
+	if err != nil {
+		return responseBody, err
+	}
 	plan.IndexPattern = indexPattern
 	plan.StartTime = startTime
 	plan.Name = model.MainExecutionPlan
@@ -575,7 +621,6 @@ func (q *QueryRunner) handleSearchCommon(ctx context.Context, indexPattern strin
 	if decision.EnableABTesting {
 		return q.executeABTesting(ctx, plan, queryTranslator, table, body, optAsync, decision, indexPattern)
 	}
-
 	return q.executePlan(ctx, plan, queryTranslator, table, body, optAsync, nil, true)
 
 }
@@ -585,10 +630,10 @@ func (q *QueryRunner) storeAsyncSearch(qmc diag.DebugInfoCollector, id, asyncId 
 
 	if result.err == nil {
 		okStatus := 200
-		asyncResponse := queryparser.SearchToAsyncSearchResponse(result.response, asyncId, false, &okStatus)
+		asyncResponse := elastic_query_dsl.SearchToAsyncSearchResponse(result.response, asyncId, false, &okStatus)
 		responseBody, err = asyncResponse.Marshal()
 	} else {
-		responseBody, _ = queryparser.EmptyAsyncSearchResponse(asyncId, false, 503)
+		responseBody, _ = elastic_query_dsl.EmptyAsyncSearchResponse(asyncId, false, 503)
 		err = result.err
 	}
 
@@ -631,11 +676,12 @@ func (q *QueryRunner) asyncQueriesCumulatedBodySize() int {
 	return size
 }
 
-func (q *QueryRunner) HandleAsyncSearchStatus(_ context.Context, id string) ([]byte, error) {
+func (q *QueryRunner) HandleAsyncSearchStatus(ctx context.Context, id string) ([]byte, error) {
+	logger.DebugWithCtx(ctx).Msgf("handling async search status for id: %s", id)
 	if _, ok := q.AsyncRequestStorage.Load(id); ok { // there IS a result in storage, so query is completed/no longer running,
-		return queryparser.EmptyAsyncSearchStatusResponse(id, false, false, 200)
+		return elastic_query_dsl.EmptyAsyncSearchStatusResponse(id, false, false, 200)
 	} else { // there is no result so query is might be(*) running
-		return queryparser.EmptyAsyncSearchStatusResponse(id, true, true, 0) // 0 is a placeholder for missing completion status
+		return elastic_query_dsl.EmptyAsyncSearchStatusResponse(id, true, true, 0) // 0 is a placeholder for missing completion status
 	}
 	// (*) - it is an oversimplification as we're responding with "still running" status even for queries that might not exist.
 	// However since you're referring to async ID given from Quesma, we naively assume it *does* exist.
@@ -644,13 +690,13 @@ func (q *QueryRunner) HandleAsyncSearchStatus(_ context.Context, id string) ([]b
 func (q *QueryRunner) HandlePartialAsyncSearch(ctx context.Context, id string) ([]byte, error) {
 	if !strings.Contains(id, tracing.AsyncIdPrefix) {
 		logger.ErrorWithCtx(ctx).Msgf("non quesma async id: %v", id)
-		return queryparser.EmptyAsyncSearchResponse(id, false, 503)
+		return elastic_query_dsl.EmptyAsyncSearchResponse(id, false, 503)
 	}
 	if result, ok := q.AsyncRequestStorage.Load(id); ok {
 		if err := result.GetErr(); err != nil {
 			q.AsyncRequestStorage.Delete(id)
 			logger.ErrorWithCtx(ctx).Msgf("error processing async query: %v", err)
-			return queryparser.EmptyAsyncSearchResponse(id, false, 503)
+			return elastic_query_dsl.EmptyAsyncSearchResponse(id, false, 503)
 		}
 		q.AsyncRequestStorage.Delete(id)
 		// We use zstd to conserve memory, as we have a lot of async queries
@@ -672,7 +718,7 @@ func (q *QueryRunner) HandlePartialAsyncSearch(ctx context.Context, id string) (
 	} else {
 		const isPartial = true
 		logger.InfoWithCtx(ctx).Msgf("async query id : %s partial result", id)
-		return queryparser.EmptyAsyncSearchResponse(id, isPartial, 200)
+		return elastic_query_dsl.EmptyAsyncSearchResponse(id, isPartial, 200)
 	}
 }
 
