@@ -1378,7 +1378,7 @@ var AggregationTests = []AggregationTestCase{
 					Cols: []model.QueryResultCol{
 						model.NewQueryResultCol("aggr__origins__parent_count", uint64(13014)),
 						model.NewQueryResultCol("aggr__origins__key_0", "DLH"),
-						model.NewQueryResultCol("aggr__origins__count", int64(15)),
+						model.NewQueryResultCol("aggr__origins__count", int64(25)),
 						model.NewQueryResultCol("top_hits__origins__originLocation_col_0", map[string]interface{}{
 							"lat": "46.84209824",
 							"lon": "-92.19360352",
@@ -4985,7 +4985,10 @@ var AggregationTests = []AggregationTestCase{
 						},
 						{
 							"doc_count": 0,
-							"key": 2000.0
+							"key": 2000.0,
+							"2": {
+								"buckets": []
+							}
 						},
 						{
 							"doc_count": 1,
@@ -5339,6 +5342,9 @@ var AggregationTests = []AggregationTestCase{
 							"doc_count": 1410,
 							"1": {
 								"value": null
+							},
+							"2": {
+								"buckets": []
 							}
 						},
 						{
@@ -6187,7 +6193,7 @@ var AggregationTests = []AggregationTestCase{
 			],
 			"track_total_hits": true
 		}`,
-		ExpectedResponse:       `{"response": {"aggregations":{}}}`,
+		ExpectedResponse:       `{"response": {"aggregations":{"0": {"buckets": []}}}}`,
 		ExpectedPancakeResults: make([]model.QueryResultRow, 0),
 		ExpectedPancakeSQL: `
 			SELECT "aggr__0__parent_count", "aggr__0__key_0", "aggr__0__count",
@@ -6294,7 +6300,7 @@ var AggregationTests = []AggregationTestCase{
 			],
 			"track_total_hits": false
 		}`,
-		ExpectedResponse:       `{"response": {"aggregations":{}}}`,
+		ExpectedResponse:       `{"response": {"aggregations":{"0": {"buckets": []}}}}`,
 		ExpectedPancakeResults: make([]model.QueryResultRow, 0),
 		ExpectedPancakeSQL: `
 			SELECT "aggr__0__parent_count", "aggr__0__key_0", "aggr__0__count",
@@ -6398,7 +6404,7 @@ var AggregationTests = []AggregationTestCase{
 			],
 			"track_total_hits": true
 		}`,
-		ExpectedResponse:       `{"response": {"aggregations":{}}}`,
+		ExpectedResponse:       `{"response": {"aggregations":{"0": {"buckets": []}}}}`,
 		ExpectedPancakeResults: make([]model.QueryResultRow, 0),
 		ExpectedPancakeSQL: `
 			SELECT "aggr__0__parent_count", "aggr__0__key_0", "aggr__0__count",
@@ -6504,7 +6510,8 @@ var AggregationTests = []AggregationTestCase{
 							"intervalString": "3600s",
 							"seriesId":       "61ca57f1-469d-11e7-af02-69e470af7417",
 							"timeField":      "timestamp"
-						}
+						},
+						"buckets": []
 					}
 				}
 			}
@@ -6605,7 +6612,7 @@ var AggregationTests = []AggregationTestCase{
 			],
 			"track_total_hits": false
 		}`,
-		ExpectedResponse:       `{"response": {"aggregations":{}}}`,
+		ExpectedResponse:       `{"response": {"aggregations":{"0": {"buckets": []}}}}`,
 		ExpectedPancakeResults: make([]model.QueryResultRow, 0),
 		ExpectedPancakeSQL: `
 			SELECT "aggr__0__parent_count", "aggr__0__key_0", "aggr__0__count",
@@ -6831,7 +6838,7 @@ var AggregationTests = []AggregationTestCase{
 				}
 			}
 		}`,
-		ExpectedResponse:       `{"aggregations": {}}`,
+		ExpectedResponse:       `{"aggregations": {"2": {"buckets": []}}}`,
 		ExpectedPancakeResults: make([]model.QueryResultRow, 0),
 		ExpectedPancakeSQL: `
 			SELECT sum(count(*)) OVER () AS "aggr__2__parent_count",
@@ -6893,7 +6900,7 @@ var AggregationTests = []AggregationTestCase{
 			],
 			"track_total_hits": true
 		}`,
-		ExpectedResponse:       `{"response": {"aggregations":{}}}`,
+		ExpectedResponse:       `{"response": {"aggregations":{"0": {"buckets": []}}}}`,
 		ExpectedPancakeResults: make([]model.QueryResultRow, 0),
 		ExpectedPancakeSQL: `
 			SELECT "aggr__0__parent_count", "aggr__0__key_0", "aggr__0__count",
