@@ -14,7 +14,7 @@ func selectTotalCnt() string {
 	return fmt.Sprintf(`SELECT count(*) AS "column_0" FROM %s`, TableName)
 }
 func selectStar(limit int) string {
-	return fmt.Sprintf("SELECT \"message\" FROM %s LIMIT %d", TableName, limit)
+	return fmt.Sprintf("SELECT \"@timestamp\", \"message\" FROM %s LIMIT %d", TableName, limit)
 }
 
 func resultCount(cnt int) []model.QueryResultRow {
@@ -27,7 +27,10 @@ func resultSelect(cnt int) []model.QueryResultRow {
 	result := make([]model.QueryResultRow, cnt)
 	for i := range cnt {
 		result[i] = model.QueryResultRow{
-			Cols: []model.QueryResultCol{model.NewQueryResultCol("message", "example")},
+			Cols: []model.QueryResultCol{
+				model.NewQueryResultCol("@timestamp", "2021-01-01T00:00:00.000Z"),
+				model.NewQueryResultCol("message", "example"),
+			},
 		}
 	}
 	return result
@@ -95,10 +98,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "1",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					}
 				]
@@ -136,10 +141,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "1",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					},
 					{
@@ -147,10 +154,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "2",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					}
 				]
@@ -187,10 +196,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "1",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					},
 					{
@@ -198,10 +209,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "2",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					}
 				]
@@ -238,10 +251,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "1",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					},
 					{
@@ -249,10 +264,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "2",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					}
 				]
@@ -289,10 +306,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "1",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					},
 					{
@@ -300,10 +319,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "2",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					}
 				]
@@ -340,10 +361,12 @@ var FullSearchRequests = []FullSearchTestCase{
 						"_id": "1",
 						"_score": 0.0,
 						"_source": {
-							"message": "example"
+							"message": "example",
+							"@timestamp": "2021-01-01T00:00:00.000Z"
 						},
 						"fields": {
-							"message": ["example"]
+							"message": ["example"],
+							"@timestamp": ["2021-01-01T00:00:00.000Z"]
 						}
 					}
 				]
@@ -355,4 +378,121 @@ var FullSearchRequests = []FullSearchTestCase{
 
 	// HitsInfo == ...
 
+	{ // [7]
+		Name: "Turing regression test",
+		QueryRequestJson: `
+		{
+			"aggs": {
+				"2": {
+					"aggs": {
+						"3": {
+							"terms": {
+								"field": "score",
+								"order": {
+									"_count": "desc"
+								},
+								"shard_size": 25,
+								"size": 5
+							}
+						}
+					},
+					"date_histogram": {
+						"field": "@timestamp",
+						"fixed_interval": "12h",
+						"min_doc_count": 1,
+						"time_zone": "Europe/Warsaw"
+					}
+				}
+			},
+			"size": 0
+		}`,
+		ExpectedResponse: `
+		{
+			"aggregations": {
+				"2": {
+					"buckets": []
+				}
+			},
+			"hits": {
+				"total": {
+					"value": 0,
+					"relation": "gte"
+				},
+				"max_score": null,
+				"hits": []
+			}
+		}`,
+		ExpectedSQLs: []string{
+			`SELECT "metric____quesma_total_count_col_0", "aggr__2__key_0", "aggr__2__count",
+			  "aggr__2__3__parent_count", "aggr__2__3__key_0", "aggr__2__3__count"
+			FROM (
+			  SELECT "metric____quesma_total_count_col_0", "aggr__2__key_0",
+				"aggr__2__count", "aggr__2__3__parent_count", "aggr__2__3__key_0",
+				"aggr__2__3__count",
+				dense_rank() OVER (ORDER BY "aggr__2__key_0" ASC) AS "aggr__2__order_1_rank"
+				,
+				dense_rank() OVER (PARTITION BY "aggr__2__key_0" ORDER BY
+				"aggr__2__3__count" DESC, "aggr__2__3__key_0" ASC) AS
+				"aggr__2__3__order_1_rank"
+			  FROM (
+				SELECT sum(count(*)) OVER () AS "metric____quesma_total_count_col_0",
+				  toInt64((toUnixTimestamp64Milli("@timestamp")+timeZoneOffset(toTimezone(
+				  "@timestamp", 'Europe/Warsaw'))*1000) / 43200000) AS "aggr__2__key_0",
+				  sum(count(*)) OVER (PARTITION BY "aggr__2__key_0") AS "aggr__2__count",
+				  sum(count(*)) OVER (PARTITION BY "aggr__2__key_0") AS
+				  "aggr__2__3__parent_count", NULL AS "aggr__2__3__key_0",
+				  count(*) AS "aggr__2__3__count"
+				FROM __quesma_table_name
+				GROUP BY toInt64((toUnixTimestamp64Milli("@timestamp")+timeZoneOffset(
+				  toTimezone("@timestamp", 'Europe/Warsaw'))*1000) / 43200000) AS
+				  "aggr__2__key_0", NULL AS "aggr__2__3__key_0"))
+			WHERE "aggr__2__3__order_1_rank"<=6
+			ORDER BY "aggr__2__order_1_rank" ASC, "aggr__2__3__order_1_rank" ASC`,
+		},
+		ExpectedSQLResults: [][]model.QueryResultRow{{}},
+	},
+	{ // [8]
+		Name: "Turing regression test",
+		QueryRequestJson: `
+		{
+			"aggs": {
+				"2": {
+					"terms": {
+						"field": "path_id",
+						"order": {
+							"_count": "desc"
+						},
+						"size": 20
+					}
+				}
+			},
+			"size": 0
+		}`,
+		ExpectedResponse: `
+		{
+			"aggregations": {
+				"2": {
+					"buckets": []
+				}
+			},
+			"hits": {
+				"total": {
+					"value": 0,
+					"relation": "gte"
+				},
+				"max_score": null,
+				"hits": []
+			}
+		}`,
+		ExpectedSQLs: []string{
+			`SELECT sum(count(*)) OVER () AS "metric____quesma_total_count_col_0",
+			  sum(count(*)) OVER () AS "aggr__2__parent_count", NULL AS "aggr__2__key_0",
+			  count(*) AS "aggr__2__count"
+			FROM __quesma_table_name
+			GROUP BY NULL AS "aggr__2__key_0"
+			ORDER BY "aggr__2__count" DESC, "aggr__2__key_0" ASC
+			LIMIT 21`,
+		},
+		ExpectedSQLResults: [][]model.QueryResultRow{{}},
+	},
 }
