@@ -4,19 +4,28 @@ package quesma_api
 
 import (
 	"context"
+	"fmt"
 	"github.com/QuesmaOrg/quesma/quesma/v2/core/tracing"
+	"github.com/hashicorp/golang-lru/v2/expirable"
 	"github.com/rs/zerolog"
+	"hash/fnv"
 	"os"
 	"time"
 )
 
 const (
-	RID                              = "request_id" // request id key for the logger
-	Reason                           = "reason"     // Known error reason key for the logger
-	Path                             = "path"
-	AsyncId                          = "async_id"
-	OpaqueId                         = "opaque_id"
-	ReasonPrefixUnsupportedQueryType = "unsupported_search_query: " // Reason for Error messages for unsupported queries will start with this prefix
+	RID      = "request_id" // request id key for the logger
+	Reason   = "reason"     // Known error reason key for the logger
+	Path     = "path"
+	AsyncId  = "async_id"
+	OpaqueId = "opaque_id"
+
+	DefaultBurstSamplerPeriodSeconds    = 20  // burst up to 600 lines of logs per 20 seconds period
+	DefaultBurstSamplerMaxLogsPerSecond = 30  // ~100k lines of logs per hour
+	DefaultSheddingFrequency            = 100 // when the limit is exhausted, log every ~ 100 log lines
+
+	DeduplicatedLogsCacheSize  = 1000
+	DeduplicatedLogsExpiryTime = 1 * time.Minute
 )
 
 type QuesmaLogger interface {
@@ -26,6 +35,10 @@ type QuesmaLogger interface {
 	Error() *zerolog.Event
 	Fatal() *zerolog.Event
 	Panic() *zerolog.Event
+
+	// TODO: Add similar for other log levels
+	DeduplicatedInfo() DeduplicatedEvent
+	DeduplicatedWarn() DeduplicatedEvent
 
 	DebugWithCtx(ctx context.Context) *zerolog.Event
 	InfoWithCtx(ctx context.Context) *zerolog.Event
@@ -42,18 +55,19 @@ type QuesmaLogger interface {
 
 type QuesmaLoggerImpl struct {
 	zerolog.Logger
+
+	deduplicatedLogs *expirable.LRU[any, struct{}]
 }
 
 func NewQuesmaLogger(log zerolog.Logger) QuesmaLogger {
 	return &QuesmaLoggerImpl{
-		Logger: log,
+		Logger:           log,
+		deduplicatedLogs: expirable.NewLRU[any, struct{}](DeduplicatedLogsCacheSize, nil, DeduplicatedLogsExpiryTime),
 	}
 }
 
 func (l *QuesmaLoggerImpl) WithComponent(name string) QuesmaLogger {
-	return &QuesmaLoggerImpl{
-		Logger: l.Logger.With().Str("component", name).Logger(),
-	}
+	return NewQuesmaLogger(l.Logger.With().Str("component", name).Logger())
 }
 
 func (l *QuesmaLoggerImpl) MarkTraceEndWithCtx(ctx context.Context) *zerolog.Event {
@@ -121,6 +135,45 @@ func (l *QuesmaLoggerImpl) ErrorWithCtx(ctx context.Context) *zerolog.Event {
 	return event
 }
 
+func (l *QuesmaLoggerImpl) DeduplicatedInfo() DeduplicatedEvent {
+	return DeduplicatedEvent{
+		event: l.Info(),
+		l:     l,
+	}
+}
+
+func (l *QuesmaLoggerImpl) DeduplicatedWarn() DeduplicatedEvent {
+	return DeduplicatedEvent{
+		event: l.Warn(),
+		l:     l,
+	}
+}
+
+type DeduplicatedEvent struct {
+	event *zerolog.Event
+	l     *QuesmaLoggerImpl
+}
+
+func hashMsgf(format string, v ...interface{}) uint32 {
+	// []interface{} is not hashable, so we need to hash it manually
+	// For the convenience sake we just hash a Print representation
+	h := fnv.New32a()
+	fmt.Fprint(h, format, v)
+	return h.Sum32()
+}
+
+// TODO: Add wrappers for other *zerolog.Event methods
+func (m DeduplicatedEvent) Msgf(format string, v ...interface{}) {
+	hash := hashMsgf(format, v)
+
+	if m.l.deduplicatedLogs.Contains(hash) {
+		return
+	}
+
+	m.l.deduplicatedLogs.Add(hash, struct{}{})
+	m.event.Msgf(format, v...)
+}
+
 func EmptyQuesmaLogger() QuesmaLogger {
 	// not so empty :D
 	return NewQuesmaLogger(zerolog.New(
@@ -129,6 +182,11 @@ func EmptyQuesmaLogger() QuesmaLogger {
 			TimeFormat: time.StampMilli,
 		}).
 		Level(zerolog.DebugLevel).
+		Sample(&zerolog.BurstSampler{
+			Burst:       DefaultBurstSamplerMaxLogsPerSecond * DefaultBurstSamplerPeriodSeconds,
+			Period:      DefaultBurstSamplerPeriodSeconds * time.Second,
+			NextSampler: zerolog.RandomSampler(DefaultSheddingFrequency),
+		}).
 		With().
 		Timestamp().
 		Logger())
